Update: 백엔드 폴더구조 정리
