+++ baemin-register/routes/api/check/index.js
@@ -1,0 +1,9 @@
+import express from 'express';
+
+import email from './email/index.js';
+
+const router = express.Router();
+
+router.use('/email', email);
+
+export default router;
