+++ baemin-register/routes/api/index.js
@@ -1,0 +1,9 @@
+import express from 'express';
+
+import check from './check/index.js';
+
+const router = express.Router();
+
+router.use('/check', check);
+
+export default router;
