+++ baemin-register/routes/auth/registerDetail.js
@@ -1,0 +1,9 @@
+import express from "express";
+
+const router = express.Router();
+
+router.get("/", function (req, res) {
+  res.render("auth/registerDetail", { title: "Register Phone" });
+});
+
+export default router;
