+++ baemin-register/routes/index.js
@@ -1,8 +1,14 @@
 import express from "express";
 
+import api from "./api/index.js";
+import auth from "./auth/index.js";
+
 const router = express.Router();
 
 const SESSION_AUTH_KEY = "auth";
+
+router.use("/api", api);
+router.use("/auth", auth);
 
 router.get("/", function (req, res, next) {
   const isAuth = req.session.hasOwnProperty(SESSION_AUTH_KEY);
